+++ src/Controllers/ProductController.php
@@ -4,20 +4,28 @@
 
     class ProductController{
         public function index(){
-            $category = new CategoryModel();
-            $params['categories'] = $category->getAllCategories();
+            if($_SESSION['dados_usuario']['nivelAcesso'] == 2){
+                $category = new CategoryModel();
+                $params['categories'] = $category->getAllCategories();
 
-            $product = new AddProductView($params);
+                $product = new AddProductView($params);
+            }else{
+                return redirect('/product','Usuário sem permissão.');
+            }
         }
 
         public function store($params){
-            $product = new ProductModel();
-            $result = $product->create($params);
+            if($_SESSION['dados_usuario']['nivelAcesso'] == 2){
+                $product = new ProductModel();
+                $result = $product->create($params);
 
-            if($result){
-                return redirect('/add/product','Produto inserido com sucesso.');
-            }else{
-                return redirect('/add/product','Algo deu errado, tente novamente.');
+                if($result){
+                    return redirect('/add/product','Produto inserido com sucesso.');
+                }else{
+                    return redirect('/add/product','Algo deu errado, tente novamente.');
+                }
+            }else{  
+                return redirect('/product','Usuário sem permissão.');
             }
         }
 
@@ -36,32 +44,44 @@
             echo json_encode($product->getProductById($params));
         }
         public function edit($params){
-            $product = new ProductModel();
-            $category = new CategoryModel();
-            $params['product'] = $product->getProductById($params);
-            $params['categories'] = $category->getAllCategories();
-            $product = new EditProductView($params);
+            if($_SESSION['dados_usuario']['nivelAcesso'] == 2){
+                $product = new ProductModel();
+                $category = new CategoryModel();
+                $params['product'] = $product->getProductById($params);
+                $params['categories'] = $category->getAllCategories();
+                $product = new EditProductView($params);
+            }else{
+                return redirect('/product','Usuário sem permissão.');
+            }
         }
 
         public function update($params){
-            $product = new ProductModel();
-            $result = $product->update($params);
+            if($_SESSION['dados_usuario']['nivelAcesso'] == 2){
+                $product = new ProductModel();
+                $result = $product->update($params);
 
-            if($result){
-                return redirect('/product', 'Produto atualizado com sucesso.');
+                if($result){
+                    return redirect('/product', 'Produto atualizado com sucesso.');
+                }else{
+                    return redirect('/product', 'Algo deu errado, tente novamente.');
+                }
             }else{
-                return redirect('/product', 'Algo deu errado, tente novamente.');
+                return redirect('/product','Usuário sem permissão.');
             }
         }
 
         public function delete($params){
-            $product = new ProductModel();
-            $result = $product->delete($params);
+            if($_SESSION['dados_usuario']['nivelAcesso'] == 2){
+                $product = new ProductModel();
+                $result = $product->delete($params);
 
-            if($result){
-                return redirect('/product', 'Produto deletado com sucesso.');
+                if($result){
+                    return redirect('/product', 'Produto deletado com sucesso.');
+                }else{
+                    return redirect('/product', 'Aldo deu errado, tente novamente.');
+                }
             }else{
-                return redirect('/product', 'Aldo deu errado, tente novamente.');
+                return redirect('/product','Usuário sem permissão.');
             }
         }
     }
