+++ src/Controllers/PromotionController.php
@@ -3,16 +3,24 @@
     require_once('src/helpers/SessionValidate.php');
     class PromotionController{
         public function index(){
-            $promotion = new AddPromotionView();
+            if($_SESSION['dados_usuario']['nivelAcesso'] == 2){
+                $promotion = new AddPromotionView();
+            }else{
+                return redirect('/promotion', 'Usuário sem permissão.');
+            }
         }
         public function store($params){
-            $promotion = new PromotionModel();
-            $result = $promotion->create($params);
+            if($_SESSION['dados_usuario']['nivelAcesso'] == 2){
+                $promotion = new PromotionModel();
+                $result = $promotion->create($params);
 
-            if($result){
-                return redirect('/promotion', 'Promoção cadastrada com sucesso.');
+                if($result){
+                    return redirect('/promotion', 'Promoção cadastrada com sucesso.');
+                }else{
+                    return redirect('/promotion', 'Aldo deu errado, tente novamente.');
+                }
             }else{
-                return redirect('/promotion', 'Aldo deu errado, tente novamente.');
+                return redirect('/promotion', 'Usuário sem permissão.');
             }
         }
         public function show(){
@@ -21,28 +29,40 @@
             $promotion = new PromotionView($params);
         }
         public function edit($params){
-            $promotion = new PromotionModel();
-            $params = $promotion->getPromotionById($params);
-            $promotion = new EditPromotionView($params);
+            if($_SESSION['dados_usuario']['nivelAcesso'] == 2){
+                $promotion = new PromotionModel();
+                $params = $promotion->getPromotionById($params);
+                $promotion = new EditPromotionView($params);
+            }else{
+                return redirect('/promotion', 'Usuário sem permissão.');
+            }
         }
         public function update($params){
-            $promotion = new PromotionModel();
-            $result = $promotion->update($params);
+            if($_SESSION['dados_usuario']['nivelAcesso'] == 2){
+                $promotion = new PromotionModel();
+                $result = $promotion->update($params);
 
-            if($result){
-                return redirect('/promotion', 'Promoção atualizada com sucesso.');
+                if($result){
+                    return redirect('/promotion', 'Promoção atualizada com sucesso.');
+                }else{
+                    return redirect('/promotion', 'Aldo deu errado, tente novamente.');
+                }
             }else{
-                return redirect('/promotion', 'Aldo deu errado, tente novamente.');
+                return redirect('/promotion', 'Usuário sem permissão.');
             }
         }
         public function delete($params){
-            $promotion = new PromotionModel();
-            $result = $promotion->delete($params);
+            if($_SESSION['dados_usuario']['nivelAcesso'] == 2){
+                $promotion = new PromotionModel();
+                $result = $promotion->delete($params);
 
-            if($result){
-                return redirect('/promotion', 'Promoção deletada com sucesso.');
+                if($result){
+                    return redirect('/promotion', 'Promoção deletada com sucesso.');
+                }else{
+                    return redirect('/promotion', 'Aldo deu errado, tente novamente.');
+                }
             }else{
-                return redirect('/promotion', 'Aldo deu errado, tente novamente.');
+                return redirect('/promotion', 'Usuário sem permissão.');
             }
         }
     }
